Relax Bundler constraint (allow 2.x)
Bundler 2.0 has been released this month.  This change relaxes
the development dependency constraint in order to allow this version.
A new constraint which forbids 3.x has been introduced instead.

Keeping the constraint unchanged, apart from being potentially
inconvenient for developers, could lead to a problem, which has been
observed in other gems developed by Ribose Inc. — some Rubies in Travis
CI come with Bundler preinstalled, causing builds to fail if it is
Bunlder 2.  An alternative solution is to enforce a specific Bundler
version in Travis (simply installing it is not enough).  However,
allowing Bundler 2 is simpler and cleaner in case of this gem.

This gem does not rely on Bundler except for managing development gems,
therefore no regression should happen.