ci: Switch to use-ruby-action in CI
The stock setup-ruby action offers only few Ruby versions, not even
the most recent ones.  Hence, it is not very suitable for testing
a Ruby gem.  In contrast, use-ruby-action, which is maintained by
well-known Ruby contributor, provides a wide range of Rubies, including
CRuby, JRuby and TruffleRuby.

Ruby versions in job's matrix must be updated because the new action
expects that in a slightly different format.

https://github.com/marketplace/actions/setup-ruby-jruby-and-truffleruby